Extract decorate request wrapper (#182)
* extract resolveOptions to a module

* extract_decorateRequestWrapper
