+++ lib/asBuffer.js
@@ -1,0 +1,32 @@
+'use strict';
+
+function asBuffer(body, options) {
+
+  var ret;
+  if (Buffer.isBuffer(body)) {
+    ret = body;
+  } else if (typeof body === 'object') {
+    ret = new Buffer(JSON.stringify(body), options.reqBodyEncoding);
+  } else if (typeof body === 'string') {
+    ret = new Buffer(body, options.reqBodyEncodeing);
+  }
+  return ret;
+}
+
+function asBufferOrString(body) {
+
+  var ret;
+  if (Buffer.isBuffer(body)) {
+    ret = body;
+  } else if (typeof body === 'object') {
+    ret = JSON.stringify(body);
+  } else if (typeof body === 'string') {
+    ret = body;
+  }
+  return ret;
+}
+
+module.exports = {
+  asBuffer: asBuffer,
+  orString: asBufferOrString
+}
